[fix] rename statecheck.js => authentication.js
