+++ src/components/admin/AdminProtectedRoute.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { checkAdminState } from '../../utils/stateCheck';
+import { checkAdminState } from '../../utils/authentication';
 
 const AdminProtectedRoute = ({ children }) => {
   const navigate = useNavigate();
