+++ src/components/admin/CreateCode.js
@@ -6,8 +6,8 @@ import Logo from "../common/Logo";
 import { Header } from "../common/Header";
 import { startAttendCheck, endAttendCheck } from "../../utils/admin";
 import { COLORS } from "../../utils/theme";
-import { checkAttendStart } from "../../utils/stateCheck";
-import { checkAdminState } from "../../utils/stateCheck";
+import { checkAttendStart } from "../../utils/authentication";
+import { checkAdminState } from "../../utils/authentication";
 
 const CreateCode = () => {
   const navigate = useNavigate();
