+++ src/components/admin/SessionList.js
@@ -7,7 +7,7 @@ import { Header } from "../common/Header";
 import Logo from "../common/Logo";
 import { StyledText } from "../common/Text";
 import { formatDate } from "../../utils";
-import { checkAdminState } from "../../utils/stateCheck";
+import { checkAdminState } from "../../utils/authentication";
 import { Container, InputContainer } from "../common/Container";
 
 const SessionItem = styled.div`
