+++ src/pages/AdminPage.js
@@ -4,10 +4,10 @@ import Logo from "../components/common/Logo";
 import { Container } from "../components/common/Container"; // Ensure this is the correct export
 import { Header } from "../components/common/Header"; // Ensure this is the correct export
 import CreateSessionForm from "../components/admin/CreateSession";
-import { checkAttendStart } from "../utils/stateCheck";
+import { checkAttendStart } from "../utils/authentication";
 import { MainButton } from "../components/common/Button";
 import { useNavigate } from "react-router-dom";
-import { checkAdminState } from "../utils/stateCheck";
+import { checkAdminState } from "../utils/authentication";
 
 export default function AdminPage() {
   const navigate = useNavigate();
